feat: [EL-2935] refactor the code according to the comments
